+++ python/pyqrlew/__init__.py
@@ -1,4 +1,4 @@
-from .wrappers import Dataset, Relation, Dialect, dataset_from_database
+from .wrappers import Dataset, Relation, Dialect, Strategy, dataset_from_database
 from pyqrlew import io
 from pyqrlew import tester
 from pyqrlew.typing import *