+++ src/relation.rs
@@ -1,15 +1,9 @@
 use std::{collections::HashMap, ops::Deref, str, sync::Arc};
 use pyo3::prelude::*;
 use qrlew::{
-    ast,
-    differential_privacy::DpParameters,
-    expr::Identifier,
-    privacy_unit_tracking::PrivacyUnit,
-    relation::{self, Variant},
-    synthetic_data::SyntheticData,
-    dialect_translation::{
+    ast, dialect_translation::{
         bigquery::BigQueryTranslator, mssql::MsSqlTranslator, postgresql::PostgreSqlTranslator, RelationWithTranslator
-    }
+    }, differential_privacy::DpParameters, expr::Identifier, hierarchy::Hierarchy, privacy_unit_tracking::{self, PrivacyUnit}, relation::{self, Variant}, sql::parse_expr, synthetic_data::SyntheticData
 };
 use crate::{
     dataset::Dataset,
@@ -34,6 +28,48 @@ impl Deref for Relation {
 impl Relation {
     pub fn new(relation: Arc<relation::Relation>) -> Self {
         Relation(relation)
+    }
+}
+
+#[derive(FromPyObject, Clone)]
+pub enum PrivacyUnitType<'a> {
+    Type1(Vec<(&'a str, Vec<(&'a str, &'a str, &'a str)>, &'a str)>),
+    Type2((Vec<(&'a str, Vec<(&'a str, &'a str, &'a str)>, &'a str)>, bool)),
+    Type3((Vec<(&'a str, Vec<(&'a str, &'a str, &'a str)>, &'a str, &'a str)>, bool)),
+}
+
+impl<'a> From<PrivacyUnitType<'a>> for PrivacyUnit {
+    fn from(input: PrivacyUnitType<'a>) -> Self {
+        match input {
+            PrivacyUnitType::Type1(data) => {
+                PrivacyUnit::from(data)
+            },
+            PrivacyUnitType::Type2(data) => {
+                PrivacyUnit::from(data)
+            },
+            PrivacyUnitType::Type3(data) => {
+                PrivacyUnit::from(data)
+            },
+        }
+    }
+}
+
+/// An Enum for the privacy unit tracking propagation
+/// Soft will protect only when it does not affect the meaning of the original query and fail otherwise.
+/// Hard will protect at all cost. It will succeed most of the time.
+#[pyclass]
+#[derive(Clone)]
+pub enum Strategy {
+    Soft,
+    Hard
+}
+
+impl From<Strategy> for privacy_unit_tracking::Strategy {
+    fn from(value: Strategy) -> Self {
+        match value {
+            Strategy::Soft => privacy_unit_tracking::Strategy::Soft,
+            Strategy::Hard => privacy_unit_tracking::Strategy::Hard,
+        }
     }
 }
 
@@ -99,11 +135,12 @@ impl Relation {
     pub fn rewrite_as_privacy_unit_preserving<'a>(
         &'a self,
         dataset: &'a Dataset,
-        privacy_unit: Vec<(&'a str, Vec<(&'a str, &'a str, &'a str)>, &'a str)>,
+        privacy_unit: PrivacyUnitType<'a>,
         epsilon_delta: HashMap<&'a str, f64>,
         max_multiplicity: Option<f64>,
         max_multiplicity_share: Option<f64>,
         synthetic_data: Option<Vec<(Vec<&'a str>, Vec<&'a str>)>>,
+        strategy: Option<Strategy>
     ) -> Result<RelationWithDpEvent> {
         let relation = self.deref().clone();
         let relations = dataset.deref().relations();
@@ -136,6 +173,7 @@ impl Relation {
             synthetic_data,
             privacy_unit,
             dp_parameters,
+            strategy.map(|s| s.into())
         )?;
         Ok(RelationWithDpEvent::new(Arc::new(
             relation_with_dp_event,
@@ -165,7 +203,7 @@ impl Relation {
     pub fn rewrite_with_differential_privacy<'a>(
         &'a self,
         dataset: &'a Dataset,
-        privacy_unit: Vec<(&'a str, Vec<(&'a str, &'a str, &'a str)>, &'a str)>,
+        privacy_unit: PrivacyUnitType<'a>,
         epsilon_delta: HashMap<&'a str, f64>,
         max_multiplicity: Option<f64>,
         max_multiplicity_share: Option<f64>,
@@ -225,6 +263,33 @@ impl Relation {
             Dialect::BigQuery => ast::Query::from(RelationWithTranslator(&relation, BigQueryTranslator)).to_string(),
         }
     }
+
+    pub fn rename_fields(&self, fields: Vec<(&str, &str)>) -> Result<Self>{
+        let fields_mapping: HashMap<&str, &str> = fields.into_iter().collect();
+        let relation = self.deref().clone();
+        Ok(
+            Relation::new(
+                Arc::new(
+                    relation.rename_fields(|n, _| fields_mapping.get(n).cloned().unwrap_or("unknown").to_string())
+                )
+            )
+        )
+    }
+
+    pub fn compose(&self, relations: Vec<(Vec<String>, Relation)>) -> Result<Self> {
+        let outer_relations = self.deref();
+        let inner_relations: Hierarchy<Arc<qrlew::Relation>> = relations
+            .into_iter()
+            .map(|(path, rel)| (Identifier::from(path), rel.0))
+            .collect();
+        let composed = outer_relations.compose(&inner_relations);
+        Ok(Relation::new(Arc::new(composed)))
+    }
+
+    pub fn with_field(&self, name: &str, expr: &str) -> Result<Self> {
+        let expr = parse_expr(expr)?;
+        Ok(Relation::new(Arc::new(self.deref().clone().with_field(name, (&expr).try_into()?)))) 
+    }
 }
 
 #[cfg(test)]
@@ -234,7 +299,7 @@ mod tests {
 
     use crate::{
         dataset::Dataset,
-        relation::Relation
+        relation::{PrivacyUnitType, Relation}
     };
     use std::collections::HashMap;
 
@@ -248,8 +313,8 @@ mod tests {
         let dataset = Dataset::new(DATASET, SCHEMA, SIZE).unwrap();
         println!("{:?}", dataset.relations()[1].0) ;
 
-        let synthetic_data = Some(vec![(vec!["extract", "census"], vec!["extract", "census"])]);
-        let privacy_unit = vec![("census", Vec::<(&str, &str, &str)>::new(), "_PRIVACY_UNIT_ROW_")];
+        let synthetic_data = Some(vec![(vec!["extract", "census"], vec!["extract", "census_sd"])]);
+        let privacy_unit = PrivacyUnitType::Type1(vec![("census", Vec::<(&str, &str, &str)>::new(), "_PRIVACY_UNIT_ROW_")]);
         let budget: HashMap<&str, f64> = [("epsilon", 1.), ("delta", 0.005)].iter().cloned().collect();
 
         let queries = [
@@ -286,7 +351,6 @@ mod tests {
         let dataset = Dataset::new(DATASET, SCHEMA, SIZE).unwrap();
         println!("{:?}", dataset.relations()[1].0) ;
 
-        let tr = PostgreSqlTranslator;
         let query = r#"SELECT "age" AS s1 FROM census;"#;
         let relation = Relation::from_query(query, &dataset, None).unwrap();