+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## [0.9.19] - 2024-05-29
+### Added
+- Add strategy argument to privacy_unit_preserving [MR54](https://github.com/Qrlew/pyqrlew/pull/54)
+- Add with_field, rename_fields and compose methods to Relation. [MR54](https://github.com/Qrlew/pyqrlew/pull/54)
+
 ## [0.9.14] - 2024-03-19
 ### Fixed
 - example notebook rewrite_with_dp